:recycle: ダッシュボードコンポーネントをリファクタリングし、非同期データ取得を改善
